outsource WithMetricsFromTaskModule
