+++ src/pie_modules/models/mixins/__init__.py
@@ -1,0 +1,1 @@
+from .with_metrics_from_taskmodule import WithMetricsFromTaskModule
