+++ src/pie_modules/models/token_classification_with_seq2seq_encoder_and_crf.py
@@ -2,7 +2,6 @@ import logging
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import torch
-from pytorch_ie import AutoTaskModule
 from pytorch_ie.core import PyTorchIEModel
 from pytorch_ie.models.interface import RequiresModelNameOrPath, RequiresNumClasses
 from pytorch_lightning.utilities.types import OptimizerLRScheduler
@@ -19,7 +18,7 @@ from transformers.modeling_outputs import TokenClassifierOutput
 from typing_extensions import TypeAlias
 
 from .components.seq2seq_encoder import build_seq2seq_encoder
-from .interface import RequiresTaskmoduleConfig
+from .mixins import WithMetricsFromTaskModule
 
 ModelInputsType: TypeAlias = BatchEncoding
 ModelTargetsType: TypeAlias = LongTensor
@@ -47,7 +46,7 @@ logger = logging.getLogger(__name__)
 
 @PyTorchIEModel.register()
 class TokenClassificationModelWithSeq2SeqEncoderAndCrf(
-    PyTorchIEModel, RequiresNumClasses, RequiresModelNameOrPath, RequiresTaskmoduleConfig
+    PyTorchIEModel, RequiresNumClasses, RequiresModelNameOrPath, WithMetricsFromTaskModule
 ):
     def __init__(
         self,
@@ -118,32 +117,6 @@ class TokenClassificationModelWithSeq2SeqEncoderAndCrf(
 
         self.setup_metrics(metric_stages=metric_stages, taskmodule_config=taskmodule_config)
 
-    def setup_metrics(
-        self, metric_stages: List[str], taskmodule_config: Optional[Dict[str, Any]] = None
-    ) -> None:
-        for stage in [TRAINING, VALIDATION, TEST]:
-            setattr(self, f"metric_{stage}", None)
-        if taskmodule_config is not None:
-            taskmodule = AutoTaskModule.from_config(taskmodule_config)
-            for stage in metric_stages:
-                if stage not in [TRAINING, VALIDATION, TEST]:
-                    raise ValueError(
-                        f'metric_stages must only contain the values "{TRAINING}", "{VALIDATION}", and "{TEST}".'
-                    )
-                stage_metric = taskmodule.configure_model_metric(stage=stage)
-                if stage_metric is not None:
-                    setattr(self, f"metric_{stage}", stage_metric)
-                else:
-                    logger.warning(
-                        f"The taskmodule {taskmodule.__class__.__name__} does not define a metric for stage "
-                        f"'{stage}'."
-                    )
-        else:
-            logger.warning("No taskmodule_config was provided. Metrics will not be available.")
-
-    def get_metric(self, stage: str) -> Optional[Metric]:
-        return getattr(self, f"metric_{stage}")
-
     def decode(
         self,
         logits: FloatTensor,
@@ -212,7 +185,7 @@ class TokenClassificationModelWithSeq2SeqEncoderAndCrf(
             attentions=outputs.attentions,
         )
 
-    def step(
+    def _step(
         self,
         stage: str,
         batch: ModelStepInputType,
@@ -245,13 +218,13 @@ class TokenClassificationModelWithSeq2SeqEncoderAndCrf(
         return loss
 
     def training_step(self, batch: ModelStepInputType, batch_idx: int) -> FloatTensor:
-        return self.step(stage=TRAINING, batch=batch, metric=self.get_metric(stage=TRAINING))
+        return self._step(stage=TRAINING, batch=batch, metric=self.get_metric(stage=TRAINING))
 
     def validation_step(self, batch: ModelStepInputType, batch_idx: int) -> FloatTensor:
-        return self.step(stage=VALIDATION, batch=batch, metric=self.get_metric(stage=VALIDATION))
+        return self._step(stage=VALIDATION, batch=batch, metric=self.get_metric(stage=VALIDATION))
 
     def test_step(self, batch: ModelStepInputType, batch_idx: int) -> FloatTensor:
-        return self.step(stage=TEST, batch=batch, metric=self.get_metric(stage=TEST))
+        return self._step(stage=TEST, batch=batch, metric=self.get_metric(stage=TEST))
 
     def _on_epoch_end(self, stage: str, metric: Optional[Metric] = None) -> None:
         if metric is not None:
